# CAB432-Strobe-Server-Py-1.0.0/insomnia/strobe-openapi.yaml
openapi: 3.0.3
info:
  title: Strobe Local API
  version: 1.0.0
  contact:
    name: CAB432 Teaching Team
    email: cab432@qut.edu.au
  description: |
    Local OpenAPI specification for the Strobe Python server.

    This spec is designed for direct import into Insomnia or other API clients and reflects
    the request and response shapes implemented by the FastAPI server in python-server/src.

    Authentication uses email + password.
    Username is system-managed and mirrors the email value.

    Authentication uses JWT bearer tokens. Register or log in first, then send:

    `Authorization: Bearer <token>`
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter the JWT token only. Do not include the "Bearer " prefix.
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
        example: 550e8400-e29b-41d4-a716-446655440000
...
    QueryLimit:
      in: query
      name: limit
      required: false
      description: Maximum number of records to return. The server caps this at 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    QueryOffset:
      in: query
      name: offset
      required: false
      description: Number of records to skip before returning results.
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: ValidationError
        message:
          type: string
          example: Title is required
...
    PostAuthor:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: jake.bradford@example.com
    PostStats:
      type: object
      required: [likes, comments]
      properties:
        likes:
          type: integer
          example: 3
        comments:
          type: integer
          example: 1
    Post:
      type: object
      required:
        [
          id,
          userId,
          title,
          description,
          images,
          status,
          createdAt,
          updatedAt,
          author,
          stats,
          currentUserLiked,
        ]
      properties:
        id:
          type: string
          example: 660e8400-e29b-41d4-a716-446655440001
        userId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: Sunrise ride
        description:
          type: string
          example: Cold start, good coffee, and an empty river path.
        images:
          type: array
          items:
            type: string
          example:
            - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
        status:
          type: string
          enum: [active, hidden]
          example: active
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:10:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-08T09:10:00.000Z
        hiddenBy:
          type: string
          nullable: true
          example: null
        hiddenAt:
          type: string
          format: date-time
          nullable: true
          example: null
        author:
          $ref: "#/components/schemas/PostAuthor"
        stats:
          $ref: "#/components/schemas/PostStats"
        currentUserLiked:
          type: boolean
          example: false
...
    Moment:
      type: object
      required:
        [id, userId, imageUrl, caption, status, createdAt, expiresAt, updatedAt]
      properties:
        id:
          type: string
          example: moment_1783417870915
        userId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        imageUrl:
          type: string
          example: /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
        caption:
          type: string
          example: Afternoon lab break.
        status:
          type: string
          enum: [active, archived, hidden]
          example: active
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:30:00.000Z
        expiresAt:
          type: string
          format: date-time
          example: 2026-07-09T09:30:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-08T09:30:00.000Z
        hiddenBy:
          type: string
          nullable: true
          example: null
        hiddenAt:
          type: string
          format: date-time
          nullable: true
          example: null
        archivedAt:
          type: string
          format: date-time
          nullable: true
          example: null
...
    CreateMomentRequest:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
          example: /uploads/550e8400-e29b-41d4-a716-446655440000/post_12345678/5f9439f1.jpg
        caption:
          type: string
          example: Afternoon lab break.
...
    PostListResponse:
      type: object
      required: [message, posts]
      properties:
        message:
          type: string
          example: Success
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
...
    MomentResponse:
      type: object
      required: [message, moment]
      properties:
        message:
          type: string
          example: Created successfully
        moment:
          $ref: "#/components/schemas/Moment"
    MomentListResponse:
      type: object
      required: [message, moments]
      properties:
        message:
          type: string
          example: Success
        moments:
          type: array
          items:
            $ref: "#/components/schemas/Moment"
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Deleted successfully
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: ValidationError
            message: Title is required
    Unauthorised:
      description: Missing, invalid, or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Unauthorised
            message: Invalid or expired token
...
paths:
...
  /v1/feed:
    get:
      tags: [Feed]
      operationId: getPersonalFeed
      summary: Get personal feed
      description: Returns posts from accounts the authenticated user follows, newest first.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryOffset"
      responses:
        "200":
          description: Feed returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
              example:
                message: Success
                posts:
                  - id: 660e8400-e29b-41d4-a716-446655440011
                    userId: 550e8400-e29b-41d4-a716-446655440010
                    title: Campus coffee stop
                    description: Fast break before the next lab.
                    images: []
                    status: active
                    createdAt: 2026-07-08T09:12:00.000Z
                    updatedAt: 2026-07-08T09:12:00.000Z
                    author:
                      id: 550e8400-e29b-41d4-a716-446655440010
                      username: matthew.mckague@example.com
                    stats:
                      likes: 1
                      comments: 0
                    currentUserLiked: false
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          description: Authenticated user ID no longer exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: NotFound
                message: User not found
  /v1/moments:
    post:
      tags: [Moments]
      operationId: createMoment
      summary: Create moment
      description: Creates a new 24-hour moment for the authenticated user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMomentRequest"
            example:
              imageUrl: /uploads/550e8400-e29b-41d4-a716-446655440000/post_12345678/5f9439f1.jpg
              caption: Afternoon lab break.
      responses:
        "201":
          description: Moment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MomentResponse"
              example:
                message: Created successfully
                moment:
                  id: moment_1783417870915
                  userId: 550e8400-e29b-41d4-a716-446655440000
                  imageUrl: /uploads/550e8400-e29b-41d4-a716-446655440000/post_12345678/5f9439f1.jpg
                  caption: Afternoon lab break.
                  status: active
                  createdAt: 2026-07-08T09:30:00.000Z
                  expiresAt: 2026-07-09T09:30:00.000Z
                  updatedAt: 2026-07-08T09:30:00.000Z
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
  /v1/moments/feed:
    get:
      tags: [Moments]
      operationId: getMomentFeed
      summary: Get moment feed
      description: Returns active moments from the authenticated user and accounts they follow.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active moments returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MomentListResponse"
              example:
                message: Success
                moments:
                  - id: moment_1783417870915
                    userId: 550e8400-e29b-41d4-a716-446655440000
                    imageUrl: /uploads/550e8400-e29b-41d4-a716-446655440000/post_12345678/5f9439f1.jpg
                    caption: Afternoon lab break.
                    status: active
                    createdAt: 2026-07-08T09:30:00.000Z
                    expiresAt: 2026-07-09T09:30:00.000Z
                    updatedAt: 2026-07-08T09:30:00.000Z
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          description: Authenticated user ID no longer exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: NotFound
                message: User not found
  /v1/moments/archive:
    get:
      tags: [Moments]
      operationId: getMomentArchive
      summary: Get moment archive
      description: Returns archived moments after lifecycle normalization.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Archived moments returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MomentListResponse"
              example:
                message: Success
                moments:
                  - id: moment_1783320000000
                    userId: 550e8400-e29b-41d4-a716-446655440000
                    imageUrl: /uploads/550e8400-e29b-41d4-a716-446655440000/post_12345678/4a89ab12.jpg
                    caption: Yesterday's run.
                    status: archived
                    createdAt: 2026-07-07T08:00:00.000Z
                    expiresAt: 2026-07-08T08:00:00.000Z
                    updatedAt: 2026-07-08T08:00:10.000Z
                    archivedAt: 2026-07-08T08:00:10.000Z
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          description: Authenticated user ID no longer exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: NotFound
                message: User not found
  /v1/moments/{id}:
    delete:
      tags: [Moments]
      operationId: deleteMoment
      summary: Delete moment
      description: Deletes a moment. The owner or a moderator may perform this action.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Moment deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: Deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorised"
        "403":
          description: Authenticated user is neither the moment owner nor a moderator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Forbidden
                message: You do not have permission to delete this moment
        "404":
          description: Moment not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: NotFound
                message: Moment not found
  /v1/moments/{id}/hide:
    post:
      tags: [Moments]
      operationId: hideMoment
      summary: Hide moment
      description: Moderator-only action that marks a moment as hidden.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Moment hidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MomentResponse"
              example:
                message: Updated successfully
                moment:
                  id: moment_1783417870915
                  userId: 550e8400-e29b-41d4-a716-446655440000
                  imageUrl: /uploads/550e8400-e29b-41d4-a716-446655440000/post_12345678/5f9439f1.jpg
                  caption: Afternoon lab break.
                  status: hidden
                  createdAt: 2026-07-08T09:30:00.000Z
                  expiresAt: 2026-07-09T09:30:00.000Z
                  updatedAt: 2026-07-08T09:42:00.000Z
                  hiddenBy: 550e8400-e29b-41d4-a716-446655440999
                  hiddenAt: 2026-07-08T09:42:00.000Z
        "401":
          $ref: "#/components/responses/Unauthorised"
        "403":
          description: Caller is authenticated but not a moderator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Forbidden
                message: Moderator role required
        "404":
          description: Moment not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: NotFound
                message: Moment not found
...